backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
websockets==12.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
aiohttp==3.10.10
